Adding double quotes in makelists for fuse. also removing ; and replacing with ' ' for fixing fuse on osx.
git-svn-id: https://svn.itwm.fhg.de/svn/SDPA@5373 c3adf61f-8968-0410-94d0-83abfdf586cf

+++ trunk/fuse/fuse.cpp
@@ -279,7 +279,7 @@ static size_t copy_string ( const std::string & str
 {
   if (offset < static_cast<off_t> (str.size()))
     {
-      size = std::min (size, str.size() - offset);
+      size = std::min (size, static_cast<size_t>(str.size() - offset));
 
       std::copy (str.begin(), str.begin() + size, buf + offset);
     }
